+++ src/main/java/com/exchange_app/accounts/Account.java
@@ -3,6 +3,7 @@ package com.exchange_app.accounts;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -11,6 +12,6 @@ class Account {
     private UUID id;
     private String firstName;
     private String secondName;
-    private Double pln;
-    private Double usd;
+    private BigDecimal pln;
+    private BigDecimal usd;
 }
